Bootstrap: re-organize
